tool: split systemd into a new crate and make tool into a lib-only crate
This is necessary to make integration testing specific to the backend.
