+++ rust/tool/shared/src/esp.rs
@@ -1,20 +1,13 @@
-use std::{path::{PathBuf, Path}, array::IntoIter};
+use std::{
+    array::IntoIter,
+    path::{Path, PathBuf},
+};
 
 use anyhow::{Context, Result};
 
-use crate::common::generation::Generation;
+use crate::generation::Generation;
 
 /// Generic ESP paths which can be specific to a bootloader
-// The const generics is necessary because:
-// (a) we cannot return Trait objects in trait definitions, see https://rust-lang.github.io/impl-trait-initiative/explainer/rpit_trait.html
-// (b) we cannot return std::slice::Iter<'_, PathBuf> because PathBuf is owned
-// (c) we cannot return std::slice::Iter<'_, &PathBuf> because then Item = &&PathBuf
-// (d) we cannot return std::slice::Iter<'_, Path> because Path is unsized
-// (e) we cannot return std::slice::Iter<'_, &Path> because then Item = &&Path
-// (f) we can bring an associated type `BorrowedIterator` and then because we want to return
-// borrows, Rust needs a lifetime and the struct cannot hold the borrow so you need to bring a
-// PhantomData<'a ()> inside the structure to keep the lifetime, this is needlessly complicated
-// when this solution ensures proper monomorphisation at a small complexity cost.
 pub trait EspPaths<const N: usize> {
     /// Build an ESP path structure out of the ESP root directory
     fn new(esp: impl AsRef<Path>) -> Self;
@@ -34,7 +27,10 @@ pub struct EspGenerationPaths {
 }
 
 impl EspGenerationPaths {
-    pub fn new<const N: usize, P: EspPaths<N>>(esp_paths: &P, generation: &Generation) -> Result<Self> {
+    pub fn new<const N: usize, P: EspPaths<N>>(
+        esp_paths: &P,
+        generation: &Generation,
+    ) -> Result<Self> {
         let bootspec = &generation.spec.bootspec.bootspec;
 
         Ok(Self {
