+++ rust/tool/shared/src/generation.rs
