+++ rust/tool/shared/src/lib.rs
