+++ rust/tool/shared/src/signature.rs
