+++ rust/tool/shared/src/utils.rs
