--- rust/tool/src/systemd/mod.rs
@@ -1,6 +1,0 @@
-pub mod version;
-mod cli;
-mod esp;
-mod install;
-
-pub use cli::Cli;
