+++ rust/tool/systemd/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "lzbt-systemd"
+version = "0.3.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.71"
+stderrlog = "0.5.4"
+log = { version = "0.4.18", features = ["std"] }
+clap = { version = "4.3.1", features = ["derive"] }
+lanzaboote_tool = { path = "../shared" }
+indoc = "2.0.1"
+tempfile = "3.5.0"
+nix = { version = "0.26.2", default-features = false, features = [ "fs" ] }
+
+[dev-dependencies]
+assert_cmd = "2.0.11"
+expect-test = "1.4.1"
+filetime = "0.2.21"
+rand = "0.8.5"
