+++ rust/tool/systemd/src/main.rs
@@ -1,0 +1,11 @@
+mod cli;
+mod esp;
+mod install;
+mod version;
+
+use clap::Parser;
+use cli::Cli;
+
+fn main() {
+    Cli::parse().call(module_path!())
+}
