+++ rust/tool/systemd/src/version.rs
@@ -5,8 +5,8 @@ use std::str::FromStr;
 
 use anyhow::{Context, Result};
 
-use crate::common::os_release::OsRelease;
-use crate::common::pe;
+use lanzaboote_tool::os_release::OsRelease;
+use lanzaboote_tool::pe;
 
 /// A systemd version.
 ///
