+++ rust/uefi/rust-toolchain.toml
@@ -1,4 +1,4 @@
 [toolchain]
-channel = "1.68.2"
+channel = "1.70.0"
 components = [ "rust-src" ]
 targets = [ "x86_64-unknown-uefi" ]
